fix: cleared color should change after controlled value changed
