+++ components/color-picker/index.ts
@@ -1,6 +1,6 @@
 import ColorPicker from './ColorPicker';
 
-export type { ColorPickerProps } from './ColorPicker';
+export type { ColorPickerProps } from './interface';
 export type { Color } from './color';
 
 export default ColorPicker;
