+++ package.json
@@ -120,7 +120,7 @@
     "@ant-design/icons": "^5.3.0",
     "@ant-design/react-slick": "~1.0.2",
     "@ctrl/tinycolor": "^3.6.1",
-    "@rc-component/color-picker": "~1.5.2",
+    "@rc-component/color-picker": "~1.5.3",
     "@rc-component/mutate-observer": "^1.1.0",
     "@rc-component/tour": "~1.12.3",
     "@rc-component/trigger": "^1.18.3",
